Proof of Concept Facial detection implementation
Started code for identifying faces within the camera frame of the program. Program currently works on MacOS with Python 3.11 and in its current state is not intended for use on Raspberry Pi.

The code can label known faces, and facial encodings are saved on disk in 'face_encodings.pkl'. Pressing 'w' will trigger learning routine if there is only one face in the frame, and a prompt to label the face to learn it appears in the console.