Server-Client code split
Refactored facial detection code to split into two files, to be used separately on Client and Server devices. The Client now connects to the Server via IP and the Server should be accessible through a TCP tunnel with Ngrok.

Saving a facial embedding now functions client-side, and the client sends the frame with a designated label to the Server to be saved as a facial embedding for recognition.

Renamed some files